Python/pillow-hello/requirements.txt
bpython
pillow
